more methods

+++ Game1/src/game1/Posn.java
@@ -2,12 +2,11 @@
 package game1;
 
 public class Posn {
-    int x;
-    int y;
+    double x;
+    double y;
     //Constructor
     Posn(int x, int y){
         this.x = x;
         this.y = y;
     }
-    
 }
